+++ controllers/baseController.php
@@ -1,0 +1,10 @@
+<?php
+/*
+ * Base Controller
+ */
+class BaseController {
+    function __construct() {
+        API::log_api_request($_GET);
+    }
+}
+?>
