+++ controllers/indexController.php
@@ -1,0 +1,10 @@
+<?php
+/*
+ * Index Controller
+ */
+class indexController {
+    function GET() {
+        require_once('frontpage.php');
+    }
+}
+?>
